requirements.txt
httpx
pandas
pydantic
python-dotenv
pyyaml
build
questionary
typer